working on improvement view. css, html, js. Controller changes and DB data types

+++ src/main/java/com/github/mwolf1963/improvementtracker/models/Improvement.java
@@ -15,7 +15,7 @@ public class Improvement {
     @Id
     @Column(name="improvement_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int improvement_id;
+    private long improvement_id;
     @ManyToOne
     @JoinColumn(name= "department_id" , nullable = false)
     private Department department;
@@ -68,7 +68,7 @@ public class Improvement {
 		this.part = part;
 	}
 
-	public int getImprovement_id() {
+	public long getImprovement_id() {
         return improvement_id;
     }
